requirements.txt
flask==2.3.3
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0
nltk==3.8.1
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
spacy==3.7.2
wikipedia-api==0.6.0
python-dateutil==2.8.2
openai==2.8.2
blis==0.7.10
wasabi==1.1.2
srsly==2.4.8
thinc==8.2.1
Werkzeug==2.3.7
torch==2.0.1
torchvision==0.15.2
diffusers==0.19.3
transformers==4.31.0
accelerate==0.21.0
Pillow==10.0.0.10
